--- workspaces/extensions/.changeset/brave-eagles-fly.md
@@ -1,5 +1,0 @@
----
-'@red-hat-developer-hub/backstage-plugin-extensions': patch
----
-
-Export translations module as default for NFS auto-discovery
