+++ workspaces/extensions/packages/app-legacy/src/components/Root/Root.tsx
@@ -46,7 +46,7 @@ import SearchIcon from '@material-ui/icons/Search';
 import { MyGroupsSidebarItem } from '@backstage/plugin-org';
 import GroupIcon from '@material-ui/icons/People';
 
-import { ExtensionsIcon } from '@red-hat-developer-hub/backstage-plugin-extensions';
+import { ExtensionsIcon } from '@red-hat-developer-hub/backstage-plugin-extensions/legacy';
 
 const useSidebarLogoStyles = makeStyles({
   root: {
