+++ workspaces/extensions/plugins/extensions/report-legacy.api.md
@@ -1,0 +1,109 @@
+## API Report File for "@red-hat-developer-hub/backstage-plugin-extensions"
+
+> Do not edit this file. It is a report generated by [API Extractor](https://api-extractor.com/).
+
+```ts
+import { BackstagePlugin } from '@backstage/core-plugin-api';
+import { FunctionComponent } from 'react';
+import { IconComponent } from '@backstage/core-plugin-api';
+import { JSX as JSX_2 } from 'react/jsx-runtime';
+import { PathParams } from '@backstage/core-plugin-api';
+import { RouteRef } from '@backstage/core-plugin-api';
+import { SubRouteRef } from '@backstage/core-plugin-api';
+
+// @public (undocumented)
+export const DynamicExtensionsPluginContent: () => JSX_2.Element;
+
+// @public
+export const DynamicExtensionsPluginRouter: () => JSX_2.Element;
+
+// @public @deprecated (undocumented)
+export const DynamicMarketplacePluginContent: () => JSX_2.Element;
+
+// @public @deprecated (undocumented)
+export const DynamicMarketplacePluginRouter: () => JSX_2.Element;
+
+// @public
+export const ExtensionsFullPageRouter: () => JSX_2.Element;
+
+// @public (undocumented)
+export const ExtensionsIcon: IconComponent;
+
+// @public
+export const extensionsPlugin: BackstagePlugin<
+  {
+    rootRouteRef: RouteRef<undefined>;
+    pluginsRouteRef: SubRouteRef<undefined>;
+    pluginRouteRef: SubRouteRef<PathParams<'/plugins/:namespace/:name'>>;
+    pluginInstallRouteRef: SubRouteRef<
+      PathParams<'/plugins/:namespace/:name/install'>
+    >;
+    packagesRouteRef: SubRouteRef<undefined>;
+    packageRouteRef: SubRouteRef<PathParams<'/packages/:namespace/:name'>>;
+    packageInstallRouteRef: SubRouteRef<
+      PathParams<'/packages/:namespace/:name/install'>
+    >;
+    collectionsRouteRef: SubRouteRef<undefined>;
+    collectionRouteRef: SubRouteRef<
+      PathParams<'/collections/:namespace/:name'>
+    >;
+    catalogTabRouteRef: SubRouteRef<undefined>;
+    installedTabRouteRef: SubRouteRef<undefined>;
+    installedPackageRouteRef: SubRouteRef<
+      PathParams<'/installed-packages/:namespace/:name'>
+    >;
+  },
+  {},
+  {}
+>;
+
+// @public
+export const ExtensionsTabbedPageRouter: () => JSX_2.Element;
+
+// @public @deprecated (undocumented)
+export const InstallationContextProvider: () => null;
+
+// @public @deprecated (undocumented)
+export const MarketplaceFullPageRouter: () => JSX_2.Element;
+
+// @public @deprecated (undocumented)
+export const MarketplaceIcon: FunctionComponent<{
+  fontSize?: 'medium' | 'large' | 'small' | 'inherit';
+}>;
+
+// @public @deprecated (undocumented)
+export const marketplacePlugin: BackstagePlugin<
+  {
+    rootRouteRef: RouteRef<undefined>;
+    pluginsRouteRef: SubRouteRef<undefined>;
+    pluginRouteRef: SubRouteRef<PathParams<'/plugins/:namespace/:name'>>;
+    pluginInstallRouteRef: SubRouteRef<
+      PathParams<'/plugins/:namespace/:name/install'>
+    >;
+    packagesRouteRef: SubRouteRef<undefined>;
+    packageRouteRef: SubRouteRef<PathParams<'/packages/:namespace/:name'>>;
+    packageInstallRouteRef: SubRouteRef<
+      PathParams<'/packages/:namespace/:name/install'>
+    >;
+    collectionsRouteRef: SubRouteRef<undefined>;
+    collectionRouteRef: SubRouteRef<
+      PathParams<'/collections/:namespace/:name'>
+    >;
+    catalogTabRouteRef: SubRouteRef<undefined>;
+    installedTabRouteRef: SubRouteRef<undefined>;
+    installedPackageRouteRef: SubRouteRef<
+      PathParams<'/installed-packages/:namespace/:name'>
+    >;
+  },
+  {},
+  {}
+>;
+
+// @public @deprecated (undocumented)
+export const MarketplaceTabbedPageRouter: () => JSX_2.Element;
+
+// @public (undocumented)
+export const PluginsIcon: IconComponent;
+
+// (No @packageDocumentation comment for this package)
+```
